requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
decorator==4.4.2
distro==1.9.0
edge-tts==7.0.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
groq==0.25.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
proglog==0.1.12
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
sympy==1.14.0
tabulate==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
wheel==0.45.1
yarl==1.20.0
